+++ proxy-pump.js
@@ -5,7 +5,7 @@ var sys = require("sys"),
     path = require("path"),
     fs = require("fs");
      
-var port = parseFloat(process.argv[2]) || 8081;
+var port = parseInt(process.argv[2]) || 8081;
 
 var options = function(request) {
   var uri = url.parse(request.url);
